requirements-dev.txt
pytest==7.1.2
pytest-cov==3.0.0
black==22.6.0
boto3==1.24.36
boto3-stubs[essential]==1.24.35
mypy==0.971
flake8==4.0.1
python-dotenv
types-PyYAML